requirements.txt
fastapi
kivy
pytest
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
alembic
bcrypt<4.0